Fix progress bar for more than 16777216 rays

+++ src/lib_interface/input_output/progressionInfo.h
@@ -94,7 +94,7 @@ protected:
 private:
 	progressOperation* parent;
 	unsigned int sizeOfEstimation; //Nombre de fils estimé
-	float sizeOf;				   //Nombre de fils détruits
+	double sizeOf;				   //Nombre de fils détruits
 };
